documentation added and comments fixed
